AssignmentAnalysis (Hosted on google cloud)/requirements.txt
flask==2.3.3
flask-session==0.5.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-api-python-client==2.149.0
google-cloud-secret-manager==2.20.0
PyPDF2==3.0.1
nltk==3.8.1
gunicorn==22.0.0